backend/requirements.txt
Django==1.11
dj-database-url==0.4.2
djangorestframework==3.6.2
whitenoise==3.3.0
gunicorn==19.7.1
psycopg2==2.6.1